services/trial-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
boto3==1.34.34
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.1.0
apscheduler==3.10.4
PyJWT[crypto]==2.8.0
aioboto3==12.3.0
asyncpg==0.29.0
python-multipart==0.0.9
